Fix the reporting for BEGIN blocks
The stack frame for BEGIN blocks has PL_compiling as the calling COP; if there
are multiple BEGIN blocks in progress (es. an use()d module use()ing a second
module), both stack frames contain a reference to the same PL_compiling
address. Since the information is wrong anyway, and not useful for the
profiling, just ignore those stack frames.