TODO update for more efficient binary file encoding
I really don't think that there's an effective AND CPU&memory
efficient way of deduplicating strings in the file. Most of the tricks
we used in Sereal are not available here.

I suppose a good start would be simply supporting some streaming
compression format or simply using out-of-band, after-the-fact
compression and making the reader, for example, transparently read
zlib compressed data.