Store process/thread parent-child relationship in the profile file
It's going to be used on the aggregation side, to find the source code for eval
"" in the parent that create subroutines used in the child.

Also changes the unique file suffix into a sort of process identifier, hence
when a big file is split in multiple parts, it keeps the unique suffix and
gains a different ordinal.