Fix for applications embedding Perl and not using the trampoline runloop
If the application uses call_sv to call its callbacks, one of the callbacks
might enable the tracing and leave it enabled on exit, to trace subsequent
callbacks.

This change fixes the fact that tracing was not enabled when re-entering a
callback, and that calling disable_profile would exit (because it tried to do
the runloop switch dance by returning NULL as the next op without a caller able
to handle that).