Fix report generation for files containing #line directives
The specific bug happened when the part of the file not mapped by #line
directives had no samples: the rendering code for the files mapped by #line
directives expected to find a rendered file under the name of the main file,
but the main file had not been rendered (because it had no samples).

The fix is to force the main file to be always rendered.