Fix subtle but potentially catastrophic bug in set_profiler_state
When not using the trampoline, switch_runloop() runs an inner runloop,
potentialy spanning the duration of the whole program.

Since RETURN/RETURNOP perform an implicit PUTBACK, the function effectively
forced the stack to the same point it had when entering the function; in the
best case this caused the wrong value to be returned from Perl, in the worst
case it made the stack pointer point into freed/reallocated memory.