Work around an off-by-one bug in the section handling logic
The typical event stream seen by the reader is:

    - section start
      - one or more samples
    - section end
    - section start
      - one or more samples
    - section end
    ...

However with the current API, the section start/end gets reported in the next
sample, which means the first sample of a new section gets reported in the
previous section.

The correct solution is to fix the reader API as sketched in the TODO, but for
today I'm just faking the section start/end by returning a 0-weight empty
sample.