Reduce memory usage for pathological eval usage
If a process forks multiple times and uses eval STRING heavily in childs (e.g.
"caller(0)", "require Module"), the amount of memory used by source maps
skyrockets.

Since this will typically happen when using 'all_evals' or 'all_evals_always',
we know evals will be sequential, and can take advantage of this fact to reduce
memory consumption (approximately by a factor of 8).

This is only mitigation, the full solution is coming in a followup commit.