Unpessimize source map merging
Devel-StatProfiler-0.19-4-gc13c350 had a comment ending in
"and it should not hurt speed too much". That turned out to be wrong.

Overall speed is still better than the version before
Devel-StatProfiler-0.19-4-gc13c350, but there is no reason not to optimize it
further.

Now instead of unpacking the map on load and then repacking it on merge, it
tries to merge packed maps without unpacking it.