Garbage-collect data for long-dead processes
If a process has not got any sample in a long time (either directly or through
its children), it's almost certain it won't get any more samples, and there is
no point in keeping around all the metadata.

This might not be true if profiling is disabled at runtime, but since the
garbage collection is under control of the user (the method is not called
automatically), that should not be an issue in practice.