Use fully-qualified eval names
For frames that look like evals (with an "(eval XXX)" file name) use a
fully-qualified name including process ID in the report.

Using just "(eval XXX)" makes it impossible to find the right eval source code
after merging reports from multiple processes.

To be 100% correct, the process ID should be added to all frames, not just the
ones that look like an eval (because an eval could use a #line directive); I'm
choosing to ignore that case, as I'm ignoring other cases of using #line
directives in conflicting ways.