Generate a zoomable flame graph
Since both links and zoom use left click as the trigger, it's simpler to
generate two versions and provide a link to switch between the two.

I'm not sure rendering both versions in the same SVG (using a right click or a
modal interface to trigger the zoom) would be an improvement.