Improve recursive sub aggregation
The results are somewhat counterintuitive: inclusive time is counted once per
stack frame for each aggregation kind (subroutine, file, caller/callee) which
means that, for example, the sum of inclusive times at each call site will be
greater than the total inclusive time spent in the sub.

This might be changed in the future.