Fix one case when eval source code was not collected
When we have a BEGIN block in an eval STRING, the frame containing the eval
source code is (correctly) skipped by the check added in
cc32572d66f923c625c195a77f2d0907435519e8, which is OK except for the fact that
the source code for that frame is not saved.

This patch simply checks whether the skipped frame is an eval STRING and saves
the source code (depending on source code collection strategy).