Remove example output from the repository
I didn't think this through and was hoping that adding it here would
allow me to link to it on github; as it turns out, github (for obvious
security reasons) doesn't render random html on its site. So I moved it
to my own website. It is easy to re-generate in any case.

As an added bonus, hopefully this will also stop github from claiming
that this project is "mostly written in html".